requirements.txt
PyQt6>=6.6.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
pydicom>=2.4.0
